+++ src/main/java/org/tdl/vireo/model/repo/custom/EmailTemplateRepoCustom.java
@@ -8,9 +8,9 @@ public interface EmailTemplateRepoCustom {
     
     public EmailTemplate findByNameOverride(String name);
 
-    public void reorder(Integer src, Integer dest);
+    public void reorder(Long src, Long dest);
     
     public void sort(String column);
     
-    public void remove(Integer index);
+    public void remove(Long index);
 }
